+++ tests/unit/test_pep425tags.py
@@ -6,6 +6,15 @@ from mock import patch
 from pip._internal import pep425tags
 
 
+@pytest.mark.parametrize('file_tag, expected', [
+    (('py27', 'none', 'any'), 'py27-none-any'),
+    (('cp33', 'cp32dmu', 'linux_x86_64'), 'cp33-cp32dmu-linux_x86_64'),
+])
+def test_format_tag(file_tag, expected):
+    actual = pep425tags.format_tag(file_tag)
+    assert actual == expected
+
+
 @pytest.mark.parametrize('version_info, expected', [
     ((2,), '2'),
     ((2, 8), '28'),
